Bug Fixes
Fixes the issue with the toggle activation by checking the current mode of all manifestations.
Fixes a bug where save states would be deleted after quitting a world by saving a copy as nbt data attached to the SpiritwebCapability.
Adds a text pop-up to help show what is happening when saving and activating power save states.