New Shardplate and blade stuff
Data for Shardplate and Blades -- including order, state of living or deadeye, who their bonded to, and how long they've been trying to bond to a player -- are now controlled by capabilities.

Suits of shardplate for each order have been removed.

The color of shardplate is now determined by the capability and its data.